Move SSLTestUtils to util package
