+++ auth0/src/test/java/com/auth0/android/authentication/AuthenticationAPIClientTest.java
@@ -9,7 +9,6 @@ import com.auth0.android.request.AuthenticationRequest;
 import com.auth0.android.request.HttpMethod;
 import com.auth0.android.request.NetworkingClient;
 import com.auth0.android.request.RequestOptions;
-import com.auth0.android.request.SSLTestUtils;
 import com.auth0.android.request.ServerResponse;
 import com.auth0.android.request.internal.RequestFactory;
 import com.auth0.android.request.internal.ThreadSwitcherShadow;
@@ -20,6 +19,7 @@ import com.auth0.android.result.UserProfile;
 import com.auth0.android.util.Auth0UserAgent;
 import com.auth0.android.util.AuthenticationAPIMockServer;
 import com.auth0.android.util.MockAuthenticationCallback;
+import com.auth0.android.util.SSLTestUtils;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.reflect.TypeToken;
